multiple-acquisition: set up continuous acquisition mode
An important configuration for acquiring multiple images is to set the
acquisition mode either to `ARV_ACQUISITION_MODE_MULTI_FRAME` or
`ARV_ACQUISITION_MODE_CONTINUOUS`. This is because examples and
implementations directly based on them will stall at the first frame if
the camera was externally set to `ARV_ACQUISITION_MODE_SINGLE_FRAME`.
This produces an unexpected behavior.

Continuous mode has been chosen here so that there is no need to
configure the number of expected frames, making the script a bit
cleaner.